pyproject.toml
[tool.poetry]
name = "Tesseract_benchmark_detections"
version = "1.0.0"
description = ""
authors = ["Kobayashi-UwU <tai242415597@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
aiofiles = "23.2.1"
altair = "5.3.0"
annotated-types = "0.6.0"
anyio = "4.3.0"
attrs = "23.2.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
contourpy = "1.2.1"
cycler = "0.12.1"
fastapi = "0.110.3"
ffmpy = "0.3.2"
filelock = "3.14.0"
fonttools = "4.51.0"
fsspec = "2024.3.1"
gradio = "4.29.0"
gradio_client = "0.16.1"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
huggingface-hub = "0.22.2"
idna = "3.7"
importlib_resources = "6.4.0"
Jinja2 = "3.1.3"
jsonschema = "4.21.1"
jsonschema-specifications = "2023.12.1"
kiwisolver = "1.4.5"
markdown-it-py = "3.0.0"
MarkupSafe = "2.1.5"
matplotlib = "3.8.4"
mdurl = "0.1.2"
numpy = "1.26.4"
opencv-python = "4.9.0.80"
orjson = "3.10.1"
packaging = "24.0"
pandas = "2.2.2"
pillow = "10.3.0"
pydantic = "2.7.1"
pydantic_core = "2.18.2"
pydub = "0.25.1"
Pygments = "2.17.2"
pyparsing = "3.1.2"
pytesseract = "0.3.10"
python-dateutil = "2.9.0.post0"
python-multipart = "0.0.9"
pytz = "2024.1"
PyYAML = "6.0.1"
referencing = "0.35.0"
requests = "2.31.0"
rich = "13.7.1"
rpds-py = "0.18.0"
ruff = "0.4.2"
semantic-version = "2.10.0"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
starlette = "0.37.2"
tomlkit = "0.12.0"
toolz = "0.12.1"
tqdm = "4.66.2"
typer = "0.12.3"
typing_extensions = "4.11.0"
tzdata = "2024.1"
urllib3 = "2.2.1"
uvicorn = "0.29.0"
websockets = "11.0.3"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
